Don't override existing package dependencies.
I was trying to update dependencies in svg-pathdata, but this is resetting it.

The same could be done with the devdependencies but I'm not sure if there are script in metapak-nfroidure which depends on specific versions, so that could break them. Just toss the PR if this would break something.